move polyline generator to gfx
